RedefStructs

Documentation for RedefStructs.

RedefStructs.replace!

RedefStructs.struct_def_name

RedefStructs.@redef

RedefStructs.replace! — Method

replace!(code::Expr, old::Symbol, new::Symbol)

Replaces every appearance of old in code with new. Modifies code and returns it.

RedefStructs.struct_def_name — Method

struct_def_name(struct_def)

Returns the name of the struct being defined in a struct ... end block.

RedefStructs.@redef — Macro

@redef [mutable] struct S [<: A]
    ...
end

Define a structure in a manner that allows redefinitions.

Examples

@redef struct S
    s::String
end

@redef mutable struct S
    n::Int
end