RedefStructs.jl

Documentation for RedefStructs.jl.

This package provides the macro @redef, which allows to create structures which are redefinable.

Example

julia> using RedefStructs

julia> @redef struct S
           s::String
       end

julia> S("Hello").s
"Hello"

julia> @redef mutable struct S
           n::Int
       end

julia> S(42).n
42

Library

RedefStructs.replace!

RedefStructs.struct_def_name

RedefStructs.@redef

RedefStructs.replace! — Method

replace!(code::Expr, old::Symbol, new::Symbol)

Replaces every appearance of old in code with new. Modifies code and returns it.

RedefStructs.struct_def_name — Method

struct_def_name(struct_def)

Returns the name of the struct being defined in a struct ... end block.

RedefStructs.@redef — Macro

@redef [mutable] struct S [<: A]
    ...
end

Define a structure in a manner that allows redefinitions.

Examples

@redef struct S
    s::String
end

@redef mutable struct S
    n::Int
end